% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchSVG-package.R
\docType{package}
\name{BatchSVG-package}
\alias{BatchSVG}
\alias{BatchSVG-package}
\title{Identify Batch-Biased Features in Spatially Variable Genes}
\description{
\code{BatchSVG} is a method to identify batch-biased spatially variable genes (SVGs) in spatial transcriptomics data. The batch variable can be defined as sample, donor sex, or other batch effects of interest. The BatchSVG method is based on the binomial deviance model (Townes et al, 2019).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christinehou11/BatchSVG}
  \item \url{https://christinehou11.github.io/BatchSVG}
  \item Report bugs at \url{https://github.com/christinehou11/BatchSVG/issues}
}

}
\author{
Christine Hou
}
\keyword{internal}
