% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasDetect.R
\name{biasDetect}
\alias{biasDetect}
\title{Biased Genes Identification}
\usage{
biasDetect(list_batch_df, nSD_dev, nSD_rank)
}
\arguments{
\item{list_batch_df}{\code{list} : The list of data frame(s) generated from
\code{featureSelection()} function. The length of the data frame list
should be at least one.}

\item{nSD_dev}{\code{integer}: A numeric vector specifying the
number of standard deviation (nSD) for each batch when analyzing the
relative change in deviance. The order of values must correspond to
the order of batches in \code{list_batch_df}.
If a single value is provided, it is applied to all batches; otherwise,
it must have the same length as \code{list_batch_df}.}

\item{nSD_rank}{\code{vector}: A numeric vector specifying the
number of standard deviation (nSD) for each batch when analyzing rank
differences. The order of values must correspond to the order of batches
in \code{list_batch_df}.
If a single value is provided, it is applied to all batches; otherwise,
it must have the same length as \code{list_batch_df}.}

\item{visual}{\code{logical}: Whether to display the detected bias genes by
visualizations of \code{deviance} and \code{rank}. Default is FALSE.
If visual = TRUE, the returned format will be two parallel plots
presenting bias genes based on both \code{nSD_dev} and \code{nSD_rank}.}
}
\value{
The output values will be the list of lists with the identified bias
genes.
}
\description{
Function to identify the bias genes based on user-selected
threshold of number of standard deviation in relative change in deviance
and rank difference.
}
\examples{
batch_df <- read.csv(
    system.file("extdata","hpc_srt_batch_df.csv",package = "BatchSVG"),
    row.names = 1, check.names = FALSE)
# data frame
bias <- biasDetect(batch_df, nSD_dev = 10, nSD_rank = 5)
bias
# plots
bias_plot <- biasDetect(batch_df, nSD_dev = 10, nSD_rank = 5, visual = TRUE)
# deviance
bias_plot$deviance
# rank
bias_plot$rank
}
