% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_nSD.R
\name{svg_nSD}
\alias{svg_nSD}
\title{SVGs Plots with Relative Change in Deviance and Rank Difference}
\usage{
svg_nSD(list_batch_df, sd_interval_dev, sd_interval_rank)
}
\arguments{
\item{list_batch_df}{\code{list} : The list of data frame(s) generated from
\code{featureSelection()} function. The length of the data frame list
should be at least one.}

\item{sd_interval_dev}{\code{vector}: A numeric vector specifying the
interval widths for standard deviation bins when analyzing the
relative change in deviance.
If a single value is provided, it is applied to all batches; otherwise,
it must have the same length as \code{list_batch_df}.}

\item{sd_interval_rank}{\code{vector}: A numeric vector specifying the
interval widths for standard deviation bins when analyzing rank
differences.
If a single value is provided, it is applied to all batches; otherwise,
it must have the same length as \code{list_batch_df}.}
}
\value{
The output values are returned as
a list of data frames containing the deviance and rank with and without
the bacth effect, the corresponding difference, the corresponding nSD,
gene, gene name, and whether the gene is outlier defined by the chosen
deviance and rank cutoff. The length of the list of data frames is equal
to the number of batch effects included.
}
\description{
Function to the spatially variable genes in relative change in
deviance and rank difference colored by the number of standard deviation.
}
\examples{

plots <- svg_nSD(sample_id_df, nSD_dev = 5, nSD_rank)

}
