% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svgs_sub4}
\alias{svgs_sub4}
\title{Filtered Spatially Variable Genes (SVGs) Subset}
\format{
A data frame with:
\describe{
\item{gene_id}{Character. Ensembl Gene ID representing each gene.}
\item{n}{Integer. The number of selected samples in which the gene ranked
within the top 2,000 spatially variable genes.}
}
}
\source{
Filtered subset from \code{res_ranks} (ranked \code{nnSVG()} results).
}
\usage{
data(svgs_sub4)
}
\description{
\code{svgs_sub4} is a subset of spatially variable genes (SVGs) derived from the
\code{res_ranks} dataset. It includes genes that were ranked within the top 2,000
in at least two of the selected samples from the original dataset.
}
\examples{
data(svgs_sub4)
}
\keyword{datasets}
