% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchSVG-package.R
\docType{package}
\name{BatchSVG-package}
\alias{BatchSVG}
\alias{BatchSVG-package}
\title{Identify Batch-Biased Features in Spatially Variable Genes}
\description{
\code{BatchSVG} is a feature-based Quality Control (QC) to identify
SVGs on spatial transcriptomics data with specific types of batch effect.
Regarding to the spatial transcriptomics data experiments, the batch can be
defined as "sample", "sex", and etc.The \code{BatchSVG} method is based on
binomial deviance model (Townes et al, 2019) and applies cutoffs based on
the number of standard deviation (nSD) of relative change in deviance and
rank difference as the data-driven thresholding approach to detect the
batch-biased outliers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christinehou11/BatchSVG}
  \item \url{https://christinehou11.github.io/BatchSVG}
  \item Report bugs at \url{https://github.com/christinehou11/BatchSVG/issues}
}

}
\author{
Christine Hou
}
\keyword{internal}
