% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasDetect.R
\name{biasDetect}
\alias{biasDetect}
\title{Biased Genes Identification}
\usage{
biasDetect(list_batch_df, nSD_dev, nSD_rank)
}
\arguments{
\item{list_batch_df}{\code{list} : The list of data frame(s) generated from
\code{featureSelection()} function. The length of the data frame list
should be at least one.}

\item{nSD_dev}{\code{integer}: A numeric vector specifying the
number of standard deviation (nSD) for each batch when analyzing the
relative change in deviance. The order of values must correspond to
the order of batches in \code{list_batch_df}.
If a single value is provided, it is applied to all batches; otherwise,
it must have the same length as \code{list_batch_df}.}

\item{nSD_rank}{\code{vector}: A numeric vector specifying the
number of standard deviation (nSD) for each batch when analyzing rank
differences. The order of values must correspond to the order of batches
in \code{list_batch_df}.
If a single value is provided, it is applied to all batches; otherwise,
it must have the same length as \code{list_batch_df}.}

\item{visual}{\code{logical}: Whether to display the detected bias genes by
visualizations of \code{deviance} and \code{rank}. Default is FALSE.
If visual = TRUE, the returned format will be two parallel plots
presenting bias genes based on both \code{nSD_dev} and \code{nSD_rank}.}
}
\value{
The named list where each element corresponds to a batch.
Each batch contains:
\itemize{
\item \strong{"Plot"}: A combined ggplot object displaying two scatter
plots—one for deviance and one for rank—with an enforced 1:1 aspect
ratio.
\item \strong{"Table"}: A data frame containing identified biased genes,
with columns indicating whether they are outliers based on
deviance or rank thresholds.
}
The order of batches in the list follows the order of \code{nSD_dev} and
\code{nSD_rank} provided by the user.
}
\description{
Function to identify the bias genes based on user-selected
threshold of number of standard deviation in relative change in deviance
and rank difference.
}
\examples{
load(system.file("extdata","list_batch_df.rda",package = "BatchSVG"))

biaGenes <- biasDetect(list_batch_df, nSD_dev = c(5,4), nSD_rank = c(4,6))

biaGenes[["sample_id"]][["Table"]] # see biased genes in table format
biaGenes[["sex"]][["Plot"]] # see biased genes in plot format
}
