% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelect.R
\name{featureSelect}
\alias{featureSelect}
\title{Feature Selection}
\usage{
featureSelect(input, batch_effects = NULL, VGs = NULL)
}
\arguments{
\item{input}{\code{SpatialExperiment}: \code{SpatialExperiment} input data.
It is assumed to have an \code{assays} slot
containing \code{counts} assay for \code{devianceFeatureSelection()} to
successfully operate and calculate the deviance and rank. The
\code{logcounts} is strongly recommended to be included in \code{assays}.
Also, the input data is assumed to set up \code{rownames(input)} as
\code{genes} and \code{gene_name} in \code{rowData(input)}. The input can
be either the raw and complete data object or the filtered data object
containing only spatially varaible genes (SVGs).}

\item{batch_effects}{\code{list}: Any batch effect (\code{slide} or
\code{subject}) based on what metadata is available within input data.
The name of the batch effect within each input object
should be specified based on different scenarios.}

\item{VGs}{\code{character}: Spatially Variable Genes (SVGs)
for \code{SpatialExperiment} object. If it is a data frame, it is assumed
to contain one column of identified variable genes with column name as
gene name/ID, e.g. "ENSG00000002330".}
}
\value{
The output values are returned as
a list of data frames containing the deviance and rank with and without
the bacth effect, the corresponding difference, the corresponding nSD,
gene, gene name, and whether the gene is outlier defined by the chosen
deviance and rank cutoff. The length of the list of data frames is equal
to the number of batch effects included.
}
\description{
Function to conduct feature selection on spatial
transcriptomics data and calculate the difference of deviance and rank
with and without the selected batch effect.
}
\examples{
library(dplyr)
library(SummarizedExperiment)
# "V11L05-333_B1","V11L05-333_D1","V11L05-335_D1","V11L05-336_A1"
load(system.file("extdata","spe_sub4.rda",package = "BatchSVG"))
# load SVGs from 4 samples
svgs_sub4 <- read.csv(
    system.file("extdata","svgs_sub4.csv", package = "BatchSVG"),
    row.names = 1, check.names = FALSE) # rank <= 2000, sample n > 1
SVGs <- svgs_sub4$gene_id

batch_df <- featureSelect(spe_sub4, 
    batch_effects = c("sample_id", "sex"), VGs = SVGs)

}
